+++ electron/webpack.config.js
@@ -126,6 +126,11 @@ const mainConfig = {
                         from: path.resolve(__dirname, '..', 'node_modules', '@cliqz', 'adblocker-electron-preload', 'dist', 'preload.cjs.js').replace(/\\/g, '/'),
                         to: path.join('preview', 'assets', 'adblocker'),
                         context: path.resolve(__dirname, '..', 'node_modules', '@cliqz', 'adblocker-electron-preload', 'dist')
+                    },
+                    {
+                        from: path.resolve(__dirname, '..', 'assets', '**', '*').replace(/\\/g, '/'),
+                        to: path.join('assets'),
+                        context: path.resolve(__dirname, '..', 'assets')
                     }
                 ]
             }